msm: mdss: fix sync fence deadlock when rotator is used
In the current implementation, whenever rotator play is called, it
forces DMA pipes to be released by calling commit on video mode ctl
paths if they are using DMA pipes. Here assumption was that DMA pipes
will either be used in block mode or line mode, all or none. But this
assumption is not valid any more where we can have a use-case where
DMA0 is used for rotator and DMA1 for video mode ctl path. In this
use-case, current implementation can lead to sync fence deadlock. Fix
this by removing obsolete assumption.

Change-Id: I2d1da9dca5d958257796d617db11d93631c98f27
CRs-Fixed: 656463
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>
Signed-off-by: tarun93 <tarunmyid@gmail.com>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_overlay.c

Signed-off-by: tarun93 <tarunmyid@gmail.com>